+++ pyproj/proj.py
@@ -45,7 +45,11 @@ class Proj(Transformer):
     """
 
     def __init__(
-        self, projparams: Any = None, preserve_units: bool = True, **kwargs
+        self,
+        projparams: Any = None,
+        preserve_units: bool = True,
+        network=None,
+        **kwargs,
     ) -> None:
         """
         A Proj class instance is initialized with proj map projection
@@ -55,12 +59,19 @@ class Proj(Transformer):
         https://proj.org/operations/projections/index.html for examples of
         key/value pairs defining different map projections.
 
+        .. versionadded:: 3.0.0 network
+
         Parameters
         ----------
         projparams: int, str, dict, pyproj.CRS
             A PROJ or WKT string, PROJ dict, EPSG integer, or a pyproj.CRS instance.
         preserve_units: bool
             If false, will ensure +units=m.
+        network: bool, optional
+            Default is None, which uses the system defaults for networking.
+            If True, it will force the use of network for grids regardless of
+            any other network setting. If False, it will force disable use of
+            network for grids regardless of any other network setting.
         **kwargs:
             PROJ projection parameters.
 
@@ -131,7 +142,9 @@ class Proj(Transformer):
             projstring = self.crs.to_proj4() or self.crs.srs
 
         self.srs = re.sub(r"\s\+?type=crs", "", projstring).strip()
-        super().__init__(_Transformer.from_pipeline(cstrencode(self.srs)))
+        super().__init__(
+            _Transformer.from_pipeline(cstrencode(self.srs), network=network)
+        )
 
     def __call__(
         self,