+++ docs/history.rst
@@ -8,6 +8,7 @@ Change Log
 * Refactor Proj to inherit from Transformer (issue #624)
 * ENH: Support obects with '__array__' method (pandas.Series, xarray.DataArray, dask.array.Array) (issue #573)
 * ENH: Added :func:`pyproj.datadir.get_user_data_dir` (pull #636)
+* ENH: Add network methods to Transformer (issue #629)
 
 2.6.1
 ~~~~~